tidy, refactor, add react router dom, upgrade to bootstrap 5
