+++ src/App.css
@@ -10,7 +10,6 @@
 .App {
   font-size: 14px;
   text-align: center;
-  height: full;
   margin-bottom: 10px;
   color: gray;
   font-family: 'B612', sans-serif;
@@ -75,13 +74,13 @@
 }
 
 /* calendar */
-#root > div > div.main.mt-4.d-flex.flex-row.justify-content-between > div.d-flex.flex-row.flex-grow-1 > div.fc.fc-media-screen.fc-direction-ltr.fc-theme-standard {
+.fc-theme-standard {
   width: 100%;
   height: 85vh;
   color: var(--dark-green);
 }
 
-#root > div > div.main.mt-4.d-flex.flex-row.justify-content-between > div.d-flex.flex-row.flex-grow-1 > div.fc.fc-media-screen.fc-direction-ltr.fc-theme-standard > div.fc-header-toolbar.fc-toolbar {
+.fc-toolbar {
   display: none;
 }
