+++ src/components/Header.js
@@ -1,0 +1,9 @@
+const Header = () => {
+  return (
+    <div className="header w-100 d-flex flex-row align-items-center justify-content-between ps-4"> 
+      <div className="app-title">OneSchedule</div>
+    </div>
+  )
+}
+
+export default Header
