+++ src/components/NavBar.css
@@ -1,0 +1,3 @@
+.wrapper {
+  width: 100px
+}
