+++ src/components/Offerings.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const Offerings = () => {
+  return (
+    <div>
+      Courses
+    </div>
+  )
+}
+
+export default Offerings
