+++ src/components/SelectedCourses.js
@@ -6,7 +6,7 @@ const SelectedCourses = ({courses, isSelected, isVisible, toggleVisibility, togg
       .filter((course) => isSelected(course.id))
       .map((course) => (
         <div className="selected-course d-flex flex-row justify-content-between align-items-center" key={course.id}> 
-          <div className="selected-info flex-grow-1 pl-3 pr-2 d-flex flex-column">
+          <div className="selected-info flex-grow-1 ps-3 pe-2 d-flex flex-column">
             <div className="heading-2">{course.course_name}</div>
             <div className="selected-instructor mt-2">{course.instructor_name} ({course.id})</div>
           </div>
